insert a pin into the corners

+++ rust-tic-tac-toe/src/utils.rs
@@ -64,20 +64,20 @@ pub fn get_pin_position_from_cursor_position(
 /// Returns the position of a pin according to its address on the table
 pub fn get_pin_position_from_address(address: u8) -> (f64, f64){
 
-    let horizontal_address: u8 = match address {
-        0 | 3 | 6 => 0,
-        1 | 4 | 7 => 1,
-        _ => 2,
+    let horizontal_address: f64 = match address {
+        0 | 3 | 6 => 0.0,
+        1 | 4 | 7 => 1.0,
+        _ => 2.0,
     };
 
-    let vertical_address: u8 = match address {
-        0 | 1 | 2 => 0,
-        3 | 4 | 5 => 1,
-        _ => 2,
+    let vertical_address: f64 = match address {
+        0 | 1 | 2 => 0.0,
+        3 | 4 | 5 => 1.0,
+        _ => 2.0,
     };
 
     (
-        (horizontal_address * 126 + 66) as f64,
-        (vertical_address * 114 + 57) as f64
+        (horizontal_address * 126.0 + 66.0),
+        (vertical_address * 114.0 + 57.0)
     )
 }
